Id types changed from Long to Integer (because of database error)

+++ src/main/java/com/codeabra/entities/Instructor.java
@@ -15,7 +15,7 @@ public class Instructor extends Person {
 
     @Id
     @GeneratedValue(strategy = GenerationType.AUTO)
-    private Long instructorId;
+    private Integer instructorId;
 
 
 }
